test user site data with jest instead of runtime code
