+++ jest.config.js
@@ -33,8 +33,9 @@ module.exports = {
   },
   setupFiles: ['./jest/stylelint-rule-test.js', './jest/polyfills.js'],
   moduleNameMapper: {
-    // Jest can't resolve CSS imports
-    '^.+\\.css$': '<rootDir>/jest/emptyModule.js',
+    // Jest can't resolve CSS or asset imports
+    '^.+\\.css|jpg|jpeg|png|svg': '<rootDir>/jest/emptyModule.js',
+
     // TODO we need to allow Jest to resolve core Webpack aliases automatically
     '@docusaurus/(browserContext|BrowserOnly|ComponentCreator|constants|docusaurusContext|ExecutionEnvironment|Head|Interpolate|isInternalUrl|Link|Noop|renderRoutes|router|Translate|use.*)':
       '@docusaurus/core/lib/client/exports/$1',
@@ -42,5 +43,6 @@ module.exports = {
     '@generated/.*': '<rootDir>/jest/emptyModule.js',
     // TODO maybe use "projects" + multiple configs if we plan to add tests to another theme?
     '@theme/(.*)': '@docusaurus/theme-classic/src/theme/$1',
+    '@site/(.*)': 'website/$1',
   },
 };
